rename/product/#35 rename folderAndClass
