+++ src/main/java/com/ghbt/ghbt_starbucks/product/Projection/IProductListByCategory.java
@@ -1,0 +1,9 @@
+package com.ghbt.ghbt_starbucks.product.Projection;
+
+
+
+public interface IProductListByCategory {
+     Long getId();
+     String getName();
+     Integer getPrice();
+}
