--- src/main/java/com/ghbt/ghbt_starbucks/product/productDto/ProductListByCategory.java
@@ -1,9 +1,0 @@
-package com.ghbt.ghbt_starbucks.product.productDto;
-
-
-
-public interface ProductListByCategory {
-     Long getId();
-     String getName();
-     Integer getPrice();
-}
