+++ src/main/java/com/ghbt/ghbt_starbucks/product/service/IProductService.java
@@ -1,8 +1,7 @@
 package com.ghbt.ghbt_starbucks.product.service;
 
 import com.ghbt.ghbt_starbucks.product.model.Product;
-import com.ghbt.ghbt_starbucks.product.productDto.ProductDto;
-import com.ghbt.ghbt_starbucks.product.productDto.ProductListByCategory;
+import com.ghbt.ghbt_starbucks.product.Projection.IProductListByCategory;
 import com.ghbt.ghbt_starbucks.product.vo.RequestProduct;
 import com.ghbt.ghbt_starbucks.product.vo.ResponseProduct;
 
@@ -13,6 +12,6 @@ public interface IProductService {
     ResponseProduct getProduct(Long id);
     List<Product> getAllProduct();
 
-    List<ProductListByCategory> getProductForCategory(String search);
+    List<IProductListByCategory> getProductForCategory(String search);
 
 }
