+++ src/main/java/com/ghbt/ghbt_starbucks/product/service/ProductServiceImpl.java
@@ -3,8 +3,7 @@ package com.ghbt.ghbt_starbucks.product.service;
 import com.ghbt.ghbt_starbucks.category.model.Category;
 import com.ghbt.ghbt_starbucks.category.repository.ICategoryRepository;
 import com.ghbt.ghbt_starbucks.product.model.Product;
-import com.ghbt.ghbt_starbucks.product.productDto.ProductDto;
-import com.ghbt.ghbt_starbucks.product.productDto.ProductListByCategory;
+import com.ghbt.ghbt_starbucks.product.Projection.IProductListByCategory;
 import com.ghbt.ghbt_starbucks.product.repository.IProductRepository;
 import com.ghbt.ghbt_starbucks.product.vo.RequestProduct;
 import com.ghbt.ghbt_starbucks.product.vo.ResponseProduct;
@@ -76,8 +75,8 @@ public class ProductServiceImpl implements IProductService{
     }
 
     @Override
-    public List<ProductListByCategory> getProductForCategory(String search) {
-        List<ProductListByCategory> productList = iProductRepository.findAllProductType(search);
+    public List<IProductListByCategory> getProductForCategory(String search) {
+        List<IProductListByCategory> productList = iProductRepository.findAllProductType(search);
         return productList;
     }
 }
